Only auto flow based on settled transfers.
Previously we'd auto flow more credit to sender after a certain number
of received transfer whether they have been settled or not. This changes
the approach to keep a separate counter of transfers that have been
settled and only grant more credit once this falls below the requested
limit. It isn't a perfect approach but it is much better than what was
there before which would effectively have the effect of providing no
flow control.