Add amqp10_msg:body_bin/1 function
To allow users to get a binary representation of the body (which may be
amqp 1.0 encoded). Currently it re-encodes the decoded value, in the
future we can optimise this by delaying the actual body decoding until
it is needed. (if it ever is).

[#159434026]

(cherry picked from commit 0fa9c30ff55f5d12897055474187d8801f551ecf)